Global Function Modify and Client Project some fixes

+++ functions/functions.js
@@ -5,9 +5,29 @@ const secrets = require('../config/secrets'),
 
 module.exports = {
     response: (res, status, response, message, data, code) => {
+        if (!res || typeof res != 'object') {
+            return console.log('\x1b[36m\x1b[31m Function Response Error: \x1b[0m', 'Response Object is required to send response');
+        }
+
+        if (!status || isNaN(status)) {
+            return console.log('\x1b[36m\x1b[31m Function Response Error: \x1b[0m', 'Status Code is required to send response. Status must be integer number');
+        }
+
+        if (!response || typeof response == 'object') {
+            return console.log('\x1b[36m\x1b[31m Function Response Error: \x1b[0m', 'Standard Response is required to send response. Response must be string');
+        }
+
+        if (!code || typeof code == 'object') {
+            return console.log('\x1b[36m\x1b[31m Function Response Error: \x1b[0m', 'Custom Response Code is required to send response');
+        }
+
+        if (message && typeof message == 'object') {
+            return console.log('\x1b[36m\x1b[31m Function Response Error: \x1b[0m', 'Response message must be string');
+        }
+
         return res.status(status).json({
-            status: status,
-            response: response,
+            status: parseInt(status),
+            response: String(response),
             message: message,
             data: data,
             code: code
@@ -108,18 +128,20 @@ module.exports = {
         return output;
     },
     unlinkFile: (fileName) => {
+        // Remove File from Storage Directory
+
+        if (!filename) { return false }
+
         var path = module.exports.storageDirectory() + '/',
             filename = fileName
         if (!fs.existsSync(path)) {
             fs.mkdirSync(path, { recursive: true })
         }
         try {
-            var finder = finder = require('findit')(path),
-                finder_flag = false;
+            var finder = finder = require('findit')(path)
             finder.on('directory', function(dir, stat, stop) {
                 var newPath = dir + '/'
                 if (fs.existsSync(newPath + filename)) {
-                    finder_flag = true;
                     try {
                         fs.unlinkSync(newPath + filename)
                     } catch {}
